+++ src/Operations/Eloquent/Store.php
@@ -8,7 +8,7 @@ use MorningTrain\Laravel\Resources\Support\Pipes\Eloquent\EnsureModelInstance;
 use MorningTrain\Laravel\Resources\Support\Pipes\Eloquent\QueryModel;
 use MorningTrain\Laravel\Resources\Support\Pipes\Eloquent\QueryToModel;
 use MorningTrain\Laravel\Resources\Support\Pipes\Eloquent\UpdateModel;
-use MorningTrain\Laravel\Resources\Support\Pipes\SetModelSuccessMessage;
+use MorningTrain\Laravel\Resources\Support\Pipes\Messages\ModelUpdatedMessage;
 use MorningTrain\Laravel\Resources\Support\Pipes\TransformToView;
 use MorningTrain\Laravel\Resources\Support\Pipes\ValidatesFields;
 
@@ -35,15 +35,9 @@ class Store extends EloquentOperation
     {
         return [
             ValidatesFields::create()->fields($this->fields),
-            UpdateModel::create()->fields($this->fields)
+            UpdateModel::create()->fields($this->fields),
+            ModelUpdatedMessage::create()
         ];
-    }
-
-    protected function afterPipes()
-    {
-        return array_merge(parent::afterPipes(), [
-            SetModelSuccessMessage::create()
-        ]);
     }
 
 }